+++ pycaruna/__init__.py
@@ -1,6 +1,7 @@
 from .authenticator import Authenticator
 from .client import CarunaPlus, TimeSpan
 from .exceptions import CarunaApiError, CarunaAuthError, CarunaError
+from .utils import customer_ids_from_user, energy_kwh, normalize_energy
 
 __all__ = [
     'Authenticator',
@@ -9,4 +10,7 @@ __all__ = [
     'CarunaError',
     'CarunaPlus',
     'TimeSpan',
+    'customer_ids_from_user',
+    'energy_kwh',
+    'normalize_energy',
 ]